Moved ErrorEntity, IdObjectLong, IdObjectUUID and EditTask to model folder
