+++ src/main/java/org/udg/trackdev/spring/configuration/CustomAccessDeniedHandler.java
@@ -6,7 +6,7 @@ import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.stereotype.Component;
 import org.udg.trackdev.spring.service.Global;
-import org.udg.trackdev.spring.entity.ErrorEntity;
+import org.udg.trackdev.spring.model.ErrorEntity;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
