+++ src/main/java/org/udg/trackdev/spring/service/DemoDataSeeder.java
@@ -5,8 +5,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.udg.trackdev.spring.configuration.UserType;
-import org.udg.trackdev.spring.controller.TaskController;
 import org.udg.trackdev.spring.entity.*;
+import org.udg.trackdev.spring.model.EditTask;
 
 import java.util.*;
 
@@ -87,19 +87,19 @@ public class DemoDataSeeder {
             Task task = taskService.createTask(backlog.getId(), "Lorem ipsum dolor sit amet", reporter.getId());
 
             if(random.nextBoolean()) {
-                TaskController.EditTask editTask = buildEditTask(users, random);
+                EditTask editTask = buildEditTask(users, random);
                 taskService.editTask(task.getId(), editTask, reporter.getId());
             }
         }
         backlog = backlogService.create(group.getId());
     }
 
-    private TaskController.EditTask buildEditTask(List<User> users, Random random) {
+    private EditTask buildEditTask(List<User> users, Random random) {
         Integer points = possibleEstimationPoints.get(random.nextInt(possibleEstimationPoints.size()));
         User assignee = users.get(random.nextInt(users.size()));
         TaskStatus status = getRandomStatus(random);
 
-        TaskController.EditTask editTask = new TaskController.EditTask();
+        EditTask editTask = new EditTask();
         editTask.assignee = assignee.getUsername();
         editTask.estimationPoints = points;
         editTask.status = status;
